% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iiasawp.R
\name{iiasawp}
\alias{iiasawp}
\title{IIASA Working Papers format}
\usage{
iiasawp(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a IIASA Working Papers. Adapted
from \href{http://intranet.iiasa.ac.at/Admin/COM/templates/wp-latex-files.zip}{http://intranet.iiasa.ac.at/Admin/COM/templates/wp-latex-files.zip}.
}
\examples{

\dontrun{
library(iiasaRmarkdown)

rmarkdown::draft(file = "MyWorkingPaper.Rmd", create_dir = TRUE, 
  template = "iiasawp", package = "iiasaRmarkdown")

rmarkdown::render(input = "MyWorkingPaper/MyWorkingPaper.Rmd")

}

}

